refactor: extract framework defaults into per-framework files + helpers
Two stacked cleanups from the audit (findings #3 and #2):

1. Extract shared pattern helpers. Every Default*Config function copied
   the same five net/http ResponseWriter patterns (WriteHeader, Write,
   Error, NotFound, Redirect), the same json.Marshal pattern, and
   near-identical json.Encode/Decode/Unmarshal patterns. Lift them into:

     netHTTPResponsePatterns()       - 5 shared net/http response entries
     jsonMarshalPattern()            - generic encoding/json.Marshal
     jsonEncodePattern(recvRegex)    - json.Encoder.Encode, recv varies
     jsonDecodeRequestPattern(recv)  - json.Decoder.Decode, recv varies
     jsonUnmarshalRequestPattern(r)  - encoding/json.Unmarshal, recv varies
     stdDefaults(status)             - the common Defaults block

   Order of patterns preserved exactly so matcher priority resolution
   stays byte-identical.

2. Split each Default*Config into its own file:
     config_chi.go, config_echo.go, config_fiber.go,
     config_gin.go, config_mux.go, config_http.go

   config.go now keeps types, the four RequestContext presets, the
   shared helpers above, and nothing else. Each per-framework file is
   70-130 LOC and self-contained.

Net effect:
  config.go      1503 -> 644  LOC (-859)
  config_chi.go    -  -> 114  LOC
  config_echo.go   -  ->  92  LOC
  config_fiber.go  -  -> 130  LOC
  config_gin.go    -  ->  89  LOC
  config_http.go   -  ->  70  LOC
  config_mux.go    -  -> 113  LOC
  total          1503 -> 1252 LOC (-251)

End-to-end output unchanged: re-running apispec against
testdata/dynamic_mount_prefix produces byte-identical paths and
components.
